fixed unit tests
added marriageBeforeDivorce method back, seperated Troy's unit tests and moved isUniqueRecordId to its own module, so that US22's test didn't have to import the gedcom_parser anymore

now all 61 unit tests will run with the following:
python -m unittest
