+++ src/gedcomError.log
@@ -4,8 +4,6 @@ ERROR: INDIVIDUAL: US03: @I12@: Birth Before Death
 ERROR: INDIVIDUAL: US03: @I13@: Birth Before Death
 ERROR: INDIVIDUAL: US22: @I51@: Record ID is not unique
 ERROR: FAMILY: US12: @I32@: Invalid Mother Age: Blanche /Johnson/ is more than 60 years older than child: Kim David /OLSEN/
-ERROR: FAMILY: US21: @F16@: Husband @I42@ is not male.
-ERROR: FAMILY: US21: @F16@: Wife @I41@ is not female.
 ERROR: FAMILY: US02: @I52@: Invalid marriage and birth dates
 ERROR: FAMILY: US02: @I46@: Invalid marriage and birth dates
 ANOMALY: INDIVIDUAL: US16: @I21@: Father's Last Name of Prince doesn't match childs's Last Name of Prince Sullivan
