+++ src/output_gedcom_tables.txt
@@ -47,7 +47,7 @@
 |    @I4@    |      Avery S       |     Johnson     |   M    | 27 Jun 1999 |  18 |  True |      NA     |                                                  NA                                                 |         NA         |
 |   @I50@    |     Vendla A       |      Hansen     |   F    | 01 Dec 1909 | 108 |  True |      NA     |                                                  NA                                                 |         NA         |
 |   @I51@    |      Vera L        |      Hansen     |   F    | 14 Jan 1905 | 113 |  True |      NA     |                                                  NA                                                 |         NA         |
-| @I51@57443 | DuplicateID VeraL  |      Hansen     |   F    | 14 Jan 1905 | 113 |  True |      NA     |                                                  NA                                                 |         NA         |
+| @I51@48186 | DuplicateID VeraL  |      Hansen     |   F    | 14 Jan 1905 | 113 |  True |      NA     |                                                  NA                                                 |         NA         |
 |   @I52@    |      James P       |      Hansen     |   M    | 22 Jun 1872 | 145 |  True |      NA     |                                 ['@I48@', '@I49@', '@I50@', '@I51@']                                |     ['@I46@']      |
 |    @I5@    |   Delaney Grace    |     Johnson     |   F    |      NA     |  -1 |  True |      NA     |                                                  NA                                                 |         NA         |
 |    @I6@    |      Marilyn       |      Hansen     |   F    | 26 Dec 1944 |  61 | False | 30 Jul 2006 |                                              ['@I21@']                                              | ['@I7@', '@I20@']  |
@@ -55,7 +55,6 @@
 |    @I8@    |     Lorraine       |     Sullivan    |   F    | 10 Feb 1981 |  37 |  True |      NA     |                                                  NA                                                 |         NA         |
 |    @I9@    |   Grant Steven     |     Sullivan    |   M    | 19 Jul 1964 |  53 |  True |      NA     |                                                  NA                                                 |         NA         |
 +------------+--------------------+-----------------+--------+-------------+-----+-------+-------------+-----------------------------------------------------------------------------------------------------+--------------------+
-
 +-------+-------------+-------------+------------+----------------------------+---------+---------------------------+-----------------------------------------------------------------------------------------------------+
 |   ID  |   Married   |   Divorced  | Husband ID |        Husband Name        | Wife ID |         Wife Name         |                                               Children                                              |
 +-------+-------------+-------------+------------+----------------------------+---------+---------------------------+-----------------------------------------------------------------------------------------------------+
@@ -97,8 +96,6 @@ ERROR: FAMILY: US12: @I27@: Invalid Mother Age: Blanche /Johnson/ is more than 6
 ERROR: FAMILY: US12: @I32@: Invalid Mother Age: Blanche /Johnson/ is more than 60 years older than child: Kim David /OLSEN/
 ERROR: FAMILY: US12: @I6@: Invalid Father Age: Warren /Hansen/ is more than 80 years older than child: Marilyn /Hansen/
 ERROR: FAMILY: US12: @I6@: Invalid Mother Age: Blanche /Johnson/ is more than 60 years older than child: Marilyn /Hansen/
-ERROR: FAMILY: US21: @F16@: Husband @I42@ is not male.
-ERROR: FAMILY: US21: @F16@: Wife @I41@ is not female.
 ERROR: GENERAL: US23: N/A: Duplicate individuals were found in the GEDCOM file.
 ERROR: INDIVIDUAL: US01: @I5@: Invalid birth date
 ERROR: INDIVIDUAL: US03: @I12@: Birth Before Death
@@ -106,3 +103,7 @@ ERROR: INDIVIDUAL: US03: @I13@: Birth Before Death
 ERROR: INDIVIDUAL: US03: @I5@: Birth Before Death
 ERROR: INDIVIDUAL: US22: @I51@: Record ID is not unique
 ERROR: INDIVIDUAL: US23: @I32@ @I3222@: Duplicate individuals. Individuals @I32@ and @I3222@ are the same person.
+ERROR: INDIVIDUAL: US27: @I12@: Age Calculation may be off, individual age is less than 0
+ERROR: INDIVIDUAL: US27: @I13@: Age Calculation may be off, individual age is less than 0
+ERROR: INDIVIDUAL: US27: @I5@: Age Calculation may be off, individual age is less than 0
+ERROR: INDIVIDUAL: US27: @I5@: Can not determine age, individual has no Birth Date
