+++ src/tests/output_gedcom_tables.txt
@@ -1,0 +1,39 @@
++-------+------------+----------+--------+-------------+-----+-------+-------------+-----------+--------------------+
+|   ID  | First Name | LastName | Gender |   Birthday  | Age | Alive |    Death    |  Children |       Spouse       |
++-------+------------+----------+--------+-------------+-----+-------+-------------+-----------+--------------------+
+| @I10@ |   CRAIG    | SCHMIDT  |   M    | 01 Jan 1957 |  61 |  True |      NA     |  ['@I9@'] |     ['@I11@']      |
+| @I11@ |   GINA     | KREIGER  |   F    | 01 Jan 1965 |  53 |  True |      NA     | ['@I12@'] | ['@I10@', '@I13@'] |
+| @I12@ |    RON     | SCHMIDT  |   M    | 01 Jan 1982 |  36 |  True |      NA     |     NA    |         NA         |
+| @I13@ |   WILD     |   BILL   |   M    | 01 Jan 1960 |  58 |  True |      NA     | ['@I12@'] |     ['@I11@']      |
+| @I14@ |  STEPHEN   |   KOSS   |   M    | 01 Jan 1931 |  34 | False | 01 Jan 1965 |  ['@I1@'] |     ['@I15@']      |
+| @I15@ |   SARAH    |   DOE    |   F    | 01 Jan 1930 |  88 |  True |      NA     |     NA    | ['@I14@', '@I16@'] |
+| @I16@ |   JOHN     | JOHNSON  |   M    | 01 Jan 1935 |  83 |  True |      NA     |     NA    |     ['@I15@']      |
+| @I17@ |  HERALD    |   DOE    |   M    | 01 Jan 1880 |  70 | False | 02 Jan 1950 | ['@I15@'] |     ['@I18@']      |
+| @I18@ |   HARAH    | SAMPSON  |   F    | 01 Jan 1890 |  70 | False | 05 Jan 1960 | ['@I15@'] |     ['@I17@']      |
+|  @I1@ |  THOMAS    |   KOSS   |   M    | 01 Nov 1961 |  51 | False | 01 Mar 2013 |  ['@I3@'] |      ['@I2@']      |
+|  @I2@ |  CHERYL    |  GILLIG  |   F    | 09 Sep 1957 |  60 |  True |      NA     |     NA    |  ['@I1@', '@I4@']  |
+|  @I3@ |   TROY     |   KOSS   |   M    | 06 Dec 1991 |  26 |  True |      NA     |     NA    |      ['@I9@']      |
+|  @I4@ |   JOHN     |  PAGANO  |   M    | 01 Jan 1950 |  67 | False | 01 Jan 2017 |  ['@I5@'] |  ['@I2@', '@I6@']  |
+|  @I5@ |  JOHNJR    |  PAGANO  |   M    | 01 Jan 1982 |  36 |  True |      NA     |     NA    |         NA         |
+|  @I6@ |    SUE     |  SMITH   |   F    | 01 Jan 1951 |  67 |  True |      NA     |  ['@I5@'] |      ['@I4@']      |
+|  @I7@ |  RAYMOND   |  GILLIG  |   M    | 03 Jul 1930 |  87 | False | 01 Sep 2017 |  ['@I2@'] |      ['@I8@']      |
+|  @I8@ |   SUSAN    |   LANG   |   F    | 01 Jan 1930 |  88 |  True |      NA     |  ['@I2@'] |      ['@I7@']      |
+|  @I9@ |  MELANIE   | SCHMIDT  |   F    | 25 Oct 1994 |  23 |  True |      NA     |     NA    |      ['@I3@']      |
++-------+------------+----------+--------+-------------+-----+-------+-------------+-----------+--------------------+
++-------+-------------+-------------+------------+------------------+---------+-------------------+-----------+
+|   ID  |   Married   |   Divorced  | Husband ID |   Husband Name   | Wife ID |     Wife Name     |  Children |
++-------+-------------+-------------+------------+------------------+---------+-------------------+-----------+
+| @F10@ | 01 Jan 1920 |      NA     |   @I17@    |   HERALD /DOE/   |  @I18@  |  HARAH /SAMPSON/  | ['@I15@'] |
+|  @F1@ | 01 Jan 1990 |      NA     |    @I1@    |  THOMAS /KOSS/   |   @I2@  |  CHERYL /GILLIG/  |  ['@I3@'] |
+|  @F2@ | 01 Jan 1960 |      NA     |   @I14@    |  STEPHEN /KOSS/  |  @I15@  |    SARAH /DOE/    |  ['@I1@'] |
+|  @F3@ | 01 Jan 1985 | 01 Jan 1989 |    @I4@    |  JOHN /PAGANO/   |   @I2@  |  CHERYL /GILLIG/  |     []    |
+|  @F4@ | 01 Jan 1940 |      NA     |    @I7@    | RAYMOND /GILLIG/ |   @I8@  |    SUSAN /LANG/   |  ['@I2@'] |
+|  @F5@ |      NA     |      NA     |    @I3@    |   TROY /KOSS/    |   @I9@  | MELANIE /SCHMIDT/ |     []    |
+|  @F6@ | 01 Jan 1980 | 01 Jan 1983 |    @I4@    |  JOHN /PAGANO/   |   @I6@  |    SUE /SMITH/    |  ['@I5@'] |
+|  @F7@ | 01 Jan 1986 | 01 Jan 1999 |   @I10@    | CRAIG /SCHMIDT/  |  @I11@  |   GINA /KREIGER/  |  ['@I9@'] |
+|  @F8@ | 01 Jan 1979 | 01 Jan 1984 |   @I13@    |   WILD /BILL/    |  @I11@  |   GINA /KREIGER/  | ['@I12@'] |
+|  @F9@ | 01 Jan 1970 |      NA     |   @I16@    |  JOHN /JOHNSON/  |  @I15@  |    SARAH /DOE/    |     []    |
++-------+-------------+-------------+------------+------------------+---------+-------------------+-----------+
+
+
+ANOMALY: INDIVIDUAL: US16: @I12@: Father's Last Name of BILL doesn't match childs's Last Name of SCHMIDT
