+++ src/tests/test_us27.py
@@ -9,6 +9,7 @@ import unittest
 from datetime import date
 
 sys.path.append('../')
+sys.argv.append("../inputA.ged")
 import individual
